Move util code for biopax export to /client #300
