+++ src/client/util/index.js
@@ -1,0 +1,4 @@
+module.exports = Object.assign(
+  {},
+  require('./file')
+);
